+++ hacs.json
@@ -1,7 +1,9 @@
 {
   "name": "Atomic Calendar Revive",
-  "content_in_root": true,
-  "filename": "atomic_calendar.js",
+  "files": [
+    "dist/atomic-calendar-revive.js",
+    "dist/atomic-calendar-revive-editor.js"
+  ],
   "country": ["GB"],
   "render_readme": true
 }
