+++ package.json
@@ -1,11 +1,11 @@
 {
   "name": "atomic-calendar",
-  "version": "0.11.1",
+  "version": "1.0.0",
   "description": "Calendar Card for Home Assistant",
-  "main": "atomic_calendar.js",
+  "main": "atomic-calendar-revive.js",
   "scripts": {
     "lint": "eslint app.js",
-    "babel": "babel atomic_calendar.js --out-file atomic_calendar.js",
+    "babel": "babel atomic-calendar-revive.js --out-file atomic-calendar-revive.js",
     "rollup": "rollup -c",
     "build": "npm run rollup  && npm run babel",
     "watch": "rollup -c --watch --config rollup-dev.config.js"
@@ -14,7 +14,7 @@
     "type": "git",
     "url": "git+https://github.com/marksie1988/atomic_calendar.git"
   },
-  "author": "Wojciech Mamak",
+  "author": "Steven Marks",
   "license": "ISC",
   "devDependencies": {
     "@babel/cli": "^7.2.3",
