+++ rollup-dev.config.js
@@ -5,7 +5,7 @@ import commonjs from 'rollup-plugin-commonjs';
 export default {
     input: 'app.js',
     output: {
-        file: 'atomic_calendar.js',
+        file: 'atomic-calendar-revive.js',
         format: 'iife'
     },
 	watch: {
